pyproject.toml
[tool.poetry]
name = "webfarmd"
version = "2.7.1"
description = "Webfarm instant site-setup daemon."
authors = ["Skylar Kelty <skylar.kelty@kcl.ac.uk>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
sci-common = "^1.2.2"
sci-portal = "^2.7.4"
pika = "^1.2.0"
requests = "^2.25.1"
python-ldap = "^3.3.1"
validators = "^0.18.2"
hvac = "0.10.7"
sentry-sdk = "^1.5.0"
cryptography = "2.9"
Jinja2 = "^3.0.3"
paramiko = "^2.9.2"
scp = "^0.14.3"
click = "8.0.4"
dnspython = "^2.3.0"

[tool.poetry.dev-dependencies]
Sphinx = "^3.4.3"
sphinx-rtd-theme = "^0.5.1"
nose = "^1.3.7"
black = "^23.1.0"
mypy = "^1"
pytest = "^6.2.2"
pytest-cov = "^2.11.1"

[[tool.poetry.source]]
name = "py-sci-common"
url = "https://gitlab.er.kcl.ac.uk/api/v4/projects/18/packages/pypi/simple"
default = false
secondary = false

[[tool.poetry.source]]
name = "py-sci-portal"
url = "https://gitlab.er.kcl.ac.uk/api/v4/projects/129/packages/pypi/simple"
default = false
secondary = false

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
webfarmd = 'webfarmd.server:run_server'
webfarm = 'webfarmd.cli:run_admin_client'
update_groups = 'webfarmd.jobs.groups:update_all_groups'
update_idp_sp_list = 'webfarmd.jobs.idp:run'
le_hook = 'webfarmd.jobs.tls:le_hook'
regenerate_webserver_config = 'webfarmd.jobs.sites:regenerate_webserver_config'
export_letsencrypt_sites = 'webfarmd.jobs.sites:export_letsencrypt_sites'
deploy_site = 'webfarmd.jobs.sites:deploy_site'
security_scan = 'webfarmd.jobs.sites:run_security_scan'
security_scans = 'webfarmd.jobs.sites:run_security_scans'
check_tls = 'webfarmd.jobs.sites:check_tls'
sync_ou = 'webfarmd.jobs.groups:sync_ou'
update_letsencrypt_certs = 'webfarmd.jobs.sites:update_letsencrypt_certs'
generate_temporary_cert = 'webfarmd.jobs.tls:generate_temporary_cert'
rebuild_docker_stacks = 'webfarmd.jobs.sites:rebuild_docker_stacks'
